% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{anbo}
\alias{anbo}
\title{Community abundance data for a desert grassland (anza borrego)}
\format{A data frame with 121 rows and 4 variables:
\describe{
  \item{row}{plot coordinate}
  \item{column}{plot coordinate}
  \item{spp}{species ID}
  \item{count}{number of individuals}
}}
\usage{
anbo
}
\description{
A dataset containing the community abundace data for plant species, as well
as the locations of plots with respect to one another
}
\keyword{datasets}

