% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{arth}
\alias{arth}
\title{Arthropod community abundance data}
\format{A data frame with 547 rows and 3 variables:
\describe{
  \item{spp}{species ID}
  \item{count}{number of individuals}
  \item{mass}{biomass}
}}
\usage{
arth
}
\description{
A dataset containing the community abundace data for individuals, as well
as their body mass.
}
\keyword{datasets}

