% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dmr_beta_table.R
\name{create_dmr_beta_table}
\alias{create_dmr_beta_table}
\title{Create iDMR beta table for study samples}
\usage{
create_dmr_beta_table(
  dmr_table,
  path_cov_files,
  id_pattern = NULL,
  min_sites = 0
)
}
\description{
input:
for each sample:
- iDMR of interest
- Bismark .cov files
}
\details{
output:
iDMR beta table for study samples
}
