% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dms_beta_table.R
\name{create_dms_beta_table}
\alias{create_dms_beta_table}
\title{Create iDMS beta table for study samples}
\usage{
create_dms_beta_table(dms_table, path_cov_files, id_pattern = NULL)
}
\description{
input:
for each sample:
- iDMS of interest
- Bismark .cov files for all study samples created in during processing, created considering reads with alpha=100 only
}
\details{
output:
iDMS beta table for study samples
}
