% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sar.R
\name{empiricalSAR}
\alias{empiricalSAR}
\title{Empirical SAR}
\usage{
empiricalSAR(spp, abund, row, col, x, y, Amin, A0, upscale = FALSE,
  EAR = FALSE)
}
\arguments{
\item{spp}{}

\item{row}{}

\item{col}{}

\item{x}{}

\item{y}{}

\item{Amin}{}

\item{A0}{}

\item{upscale}{logical. Do upscaling?}

\item{EAR}{logical. TRUE computes the endemics-area relatinship}

\item{abun}{}
}
\description{
Empirical SAR
}
\examples{
esf=meteESF(spp=anbo$spp,
             abund=anbo$count)
}
\author{
Andy Rominger <ajrominger@gmail.com>, Cory Merow
}
\references{
Harte, J. 2011. Maximum entropy and ecology: a theory of abundance, distribution, and energetics. Oxford University Press.
}

