% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_cov_alpha100.R
\name{filter_cov_alpha100}
\alias{filter_cov_alpha100}
\title{Generate filtered Bismark coverage files}
\usage{
filter_cov_alpha100(
  path_bismark2bedGraph,
  path_cpg_file,
  path_read_table,
  path_out = NULL,
  min_sites = 6,
  remove_cpg = FALSE
)
}
\arguments{
\item{path_bismark2bedGraph}{The absolute path to the `bismark2bedGraph` function file}

\item{path_cpg_file}{The absolute path to the 'CpG_context' file generated by Bismark for the analyzed sample. This file must be produced using the `bismark_methylation_extractor` function with the -`-comprehensive` option, as strand-specific methylation information is not required. The file should contain five columns reporting in the following order: sequencing read ID, methylation status, chromosomal position, genomic position, and methylation call of the CpG covered by the sequencing read. For more details, consult the Bismark manual: https://felixkrueger.github.io/Bismark/bismark/.}

\item{path_read_table}{The Read Table for the analyzed sample, created using `METER::create_read_table` function, which summarizes methylation information for each sequencing read.}

\item{path_out}{The absolute path to the output folder. If not specified, the output folder will default to the same directory as the input 'CpG_context' file.}

\item{min_sites}{An integer specifying the minimum number of CpGs a read must contain to be included in the filtered "CpG_context" file and the corresponding filtered "coverage" file (default: min_sites = 6).}

\item{remove_cpg}{A logical value; set to TRUE to delete the intermediate 'CpG_context' file after the final 'coverage' file has been generated.}
}
\value{
A "coverage" file generated using the `bismark2bedGraph` function from Bismark, created considering only reads with an alpha value of 1 and covering a number of CpG sites greater than or equal to `min_sites`.
}
\description{
Generate filtered Bismark coverage files
}
