% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnseESF.R
\name{gsneESF}
\alias{gsneESF}
\title{gsneESF}
\usage{
gsneESF(gen, spp, abund, power, G0 = NULL, S0 = NULL, N0 = NULL,
  E0 = NULL, minE)
}
\arguments{
\item{gen}{A vector of genera names}

\item{spp}{A vector of species names}

\item{abund}{A vector of abundances}

\item{power}{A vector of metabolic rates}

\item{G0}{Total number of genera}

\item{S0}{Total number of species}

\item{N0}{Total number of individuals}

\item{E0}{Total metabolic rate; defaults to N0*1e6 if not specified or 
calculated from \code{power} to allow one to fit models that 
do not depend on metabolic rates}

\item{minE}{Minimum possible metabolic rate}
}
\value{
An object of class \code{meteESF} with elements
\describe{
  \item{\code{data}}{The data used to construct the ESF}
  \item{\code{emin}}{The minimum metabolic rate used to rescale metabolic rates}
  \item{\code{La}}{Vector of Lagrange multipliers}
  \item{\code{La.info}}{Termination information from optimization procedure}
  \item{\code{state.var}}{State variables used to constrain entropy maximization}
  \item{\code{Z}}{Normalization constant for ESF}
}
}
\description{
\code{gsneESF} Calculates the ``ecosystem structure
function'' using a higher taxon (e.g. genera) as an additional constraint
}
\details{
Use follows that of \code{meteESF}
}
\author{
Andy Rominger <ajrominger@gmail.com>, Cory Merow
}
\seealso{
meteESF
}
\keyword{METE,}
\keyword{MaxEnt,}
\keyword{ecosystem}
\keyword{function}
\keyword{lagrange}
\keyword{multiplier,}
\keyword{structure}

