% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ipd.meteESF.R
\name{ipd.meteESF}
\alias{ipd.meteESF}
\title{Individual Power Distribution}
\usage{
\method{ipd}{meteESF}(esf)
}
\arguments{
\item{esf}{an object of class meteESF.}
}
\value{
an object of class meteDist
}
\description{
\code{ipd.meteESF} calculates the distribution Psi(e | N0, S0, E0),
the distribution of metabolic rates across all individuals in a commmunity
}
\details{
See examples.
}
\examples{
esf1 <- meteESF(spp=arth$spp,
               abund=arth$count,
               power=arth$mass^(.75),
               minE=min(arth$mass^(.75)))
ipd1 <- ipd.meteESF(esf1)
}
\author{
Andy Rominger <ajrominger@gmail.com>, Cory Merow
}
\references{
Harte, J. 2011. Maximum entropy and ecology: a theory of abundance, distribution, and energetics. Oxford University Press.
}
\seealso{
meteDist, sad.meteESF, metePsi
}
\keyword{METE,}
\keyword{MaxEnt,}
\keyword{ecosystem}
\keyword{function}
\keyword{lagrange}
\keyword{multiplier,}
\keyword{structure}

