% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ipd.meteESF.R
\name{ipd}
\alias{ipd}
\alias{ipd.meteESF}
\title{Individual Power Distribution}
\usage{
ipd(x)

\method{ipd}{meteESF}(esf)
}
\arguments{
\item{esf}{an object of class mete.}
}
\value{
list
}
\description{
\code{ipd.meteESF} calculates the distribution Psi(e | N0, S0, E0) (the distribution of metabolic rates across all individuals in a commmunity)
}
\details{
See examples.
}
\examples{
esf1=meteESF(spp=arth$spp,
              abund=arth$count,
              power=arth$mass^(.75),
              minE=min(arth$mass^(.75)))
ipd1=ipd(esf1)
}
\author{
Andy Rominger <ajrominger@gmail.com>, Cory Merow
}
\references{
Harte, J. 2011. Maximum entropy and ecology: a theory of abundance, distribution, and energetics. Oxford University Press.
this documentation when the user looks them up from the command
line.
}
\keyword{METE,}
\keyword{MaxEnt,}
\keyword{ecosystem}
\keyword{function}
\keyword{lagrange}
\keyword{multiplier,}
\keyword{structure}

