% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logLik.meteDist.R
\name{logLik.meteDist}
\alias{logLik.meteDist}
\title{Compute log-likelihood of a meteDist object}
\usage{
\method{logLik}{meteDist}(object, ...)
}
\arguments{
\item{object}{a \code{meteDist} object}

\item{...}{arguments to be passed}
}
\value{
object of class \code{logLik}
}
\description{
\code{logLik.meteDist} computes log-likelihood of a meteDist object
}
\details{
See Examples.
}
\examples{
esf1=meteESF(spp=arth$spp,
              abund=arth$count,
              power=arth$mass^(.75),
              minE=min(arth$mass^(.75)))
ipd1=ipd.meteESF(esf1)
logLik(ipd1)
}
\author{
Andy Rominger <ajrominger@gmail.com>, Cory Merow
}
\references{
Harte, J. 2011. Maximum entropy and ecology: a theory of abundance, distribution, and energetics. Oxford University Press.
}

