% Generated by roxygen2 (4.0.2): do not edit by hand
\name{logLikZ.meteDist}
\alias{logLikZ.meteDist}
\title{Compute log-likelihood z-score}
\usage{
logLikZ.meteDist(x, nrep, return.sim = FALSE)
}
\arguments{
\item{x}{a \code{meteDist} object}

\item{nrep}{number of simulations from the fitted METE distribution}

\item{return.sim}{logical; return the simulated liklihood values}
}
\value{
list with elements
\describe{
   \item{z}{The z-score}
   \item{obs}{log-likelihood}
   \item{sim}{\code{nrep} Simulated values}
}
}
\description{
\code{logLikZ.meteDist} computes a log-likelihood z-score
}
\details{
\code{logLikZ.meteDist} simulates from a fitted METE distribution (e.g. a species abundance distribution or individual power distribution) and calculates the likelihood of these simulated data sets. The distribution of these values is compared against the likelihood of the data to obtain a z-score.
}
\examples{
esf1=meteESF(spp=arth$spp,
              abund=arth$count,
              power=arth$mass^(.75),
              minE=min(arth$mass^(.75)))
ipd1=ipd.meteESF(esf1)
llz=logLikZ.meteDist(ipd1, nrep=100, return.sim=TRUE)
plot(density(llz$sim),xlim=range(c(llz$sim,llz$obs)),
     xlab='log(likelihood)',col='red')
abline(v=llz$obs,lty=2)
legend('top',legend=c('data','simulated'),col=c('black','red'),
      lty=c(1,1),bty='n')
}
\author{
Andy Rominger <ajrominger@gmail.com>, Cory Merow
}
\references{
Harte, J. 2011. Maximum entropy and ecology: a theory of abundance, distribution, and energetics. Oxford University Press.
}
\seealso{
mseZ.meteDist
}

