\name{makeMete}
\alias{makeMete}
\title{Title of function}
\usage{
makeMete(otu, abund, power, min.e)
}
\arguments{
  \item{arg}{description of arg blah blah blah blah}

  \item{arg}{description of arg blah blahblah blah}
}
\value{
An object of class "mete"
}
\description{
\code{function.name} what it does
}
\details{
how it works etc.
}
\examples{
#code to run
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
\keyword{manip}

