% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mete.theta}
\alias{mete.theta}
\title{Title of function}
\usage{
mete.theta(epsilon, n, la2)
}
\arguments{
\item{arg}{description of arg}

\item{arg}{description of arg}
}
\value{
- the type of object that the function returns
}
\description{
\code{function.name} what it does
}
\details{
how it works
etc.
}
\examples{
#code to run
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
\keyword{manip}

