% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/meteESF.R
\name{meteESF}
\alias{meteESF}
\title{meteESF}
\usage{
meteESF(spp, abund, power, S0 = NULL, N0 = NULL, E0 = NULL, minE)
}
\arguments{
\item{spp}{A vector of species names for each entry}

\item{abund}{A vector giving abundances of each entry}

\item{power}{A vector giving metabolic rates for each entry}

\item{S0}{Total number of species}

\item{N0}{Total number of individuals}

\item{E0}{Total metabolic rate; defaults to N0*1e6 if not specified or
calculated from \code{power} to allow one to fit models that do not depend on metabolic rates}

\item{minE}{Minimum possible metabolic rate}
}
\value{
An object of class \code{esf}
}
\description{
\code{meteESF} Calculates the ``ecosystem structure
function'' which forms the core of the Maximum Entropy Theory of
Ecology
}
\details{
Uses either data or state variables to calculate the Ecosystem Structure
Function (ESF). \code{power} need not be specified; if missing an arbitrarily
 large value is assigned to E0 (N0*1e5) such that it will minimally affect
 estimation of Lagrange multipliers. Consider using sensitivity analysis to
 confirm this assumption.
}
\examples{
esf1=meteESF(spp=arth$spp,
              abund=arth$count,
              power=arth$mass^(.75),
              minE=min(arth$mass^(.75)))
}
\author{
Andy Rominger <ajrominger@gmail.com>, Cory Merow
}
\references{
Harte, J. 2011. Maximum entropy and ecology: a theory of abundance, distribution, and energetics. Oxford University Press.
}
\keyword{METE,}
\keyword{MaxEnt,}
\keyword{ecosystem}
\keyword{function}
\keyword{lagrange}
\keyword{multiplier,}
\keyword{structure}

