% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metePi.R, R/meteSSF.R
\name{metePi}
\alias{metePi}
\title{metePi}
\usage{
metePi(n, la, n0)

metePi(n, la, n0)
}
\arguments{
\item{n}{A vector giving abundances of each entry}

\item{la}{The spatial Lagrange multiplier returned by \code{meteSSF}}

\item{n0}{Total abundance in area A0}

\item{abund}{A vector giving abundances of each entry}

\item{A}{The area at which abundances were recorded}

\item{A0}{Total study area}

\item{n0}{Total abundance in area A0}
}
\value{
An object of class \code{meteSSF} with elements
\describe{
  \item{\code{data}}{The data used to construct the SSF}
  \item{\code{La}}{Vector of Lagrange multipliers}
  \item{\code{La.info}}{Termination information from optimization procedure}
  \item{\code{state.var}}{State variables used to constrain entropy maximization}
}

a numeric vector giving the probability of each entry in \code{n}
}
\description{
\code{metePi} calculates the ``spatial structure
function'' Pi(n) (analygous to the ecosystem structure function). From
the SSF the spatial abundance distribution can be calculated.

\code{metePi} is a low level function that returns the spatial species abundance
distribution \eqn{Pi(n)} predicted by METE; vectorized in n
}
\details{
Uses either data or state variables to calculate the Spatial Structure
Function (SSF). Uses internal code to determine when computation-saving approximations
can be safely made

See Examples
}
\examples{
data(anbo)
## SSF for species `crcr' in the anbo data set
pi1 <- metePi(anbo$count[anbo.new$spp=='crcr'], A=1, A0=16)
pi1
metePi(0:10, 0.01, 100)
}
\author{
Andy Rominger <ajrominger@gmail.com>, Cory Merow

Andy Rominger <ajrominger@gmail.com>, Cory Merow
}
\references{
Harte, J. 2011. Maximum entropy and ecology: a theory of abundance, distribution, and energetics. Oxford University Press.

Harte, J. 2011. Maximum entropy and ecology: a theory of abundance, distribution, and energetics. Oxford University Press.
}
\seealso{
metePi

metePi
}
\keyword{METE,}
\keyword{MaxEnt}
\keyword{lagrange}
\keyword{multiplier,}

