\name{meteR-package}
\alias{meteR-package}
\alias{meteR}
\docType{package}
\title{
Analyses with the Maximum Entropy Theory of Ecology (METE)
}
\description{
Fits and plots macroecological patterns predicted by the Maximum Entropy Theory of Ecology (METE)
}
\details{
\tabular{ll}{
Package: \tab meteR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-01-04\cr
License: \tab GPL-2\cr
}
%~~ An overview of how to use the package, including the ~~
%~~ most important functions ~~
}
\author{
Andy Rominger, Cory Merow, John Harte

Maintainer: Cory Merow \email{cory.merow@gmail.com}
}
\references{
  Harte, J., Zillio, T., Conlisk, E. & Smith, A. (2008). Maximum entropy and the state-variable approach to macroecology. Ecology, 89, 2700–2711.
  
  Harte, J. (2008). From Spatial Pattern in the Distribution and Abundance of Species to a Unified Theory of Ecology: The Role of Maximum Entropy Methods. Applied Optimization pp. 243–272. Applied Optimization. Springer Berlin Heidelberg, Berlin, Heidelberg.
  
  Harte, J., Smith, A.B. & Storch, D. (2009). Biodiversity scales from plots to biomes with a universal species-area curve. Ecology Letters, 12, 789–797.
  
  Harte, J. (2011). Maximum entropy and ecology: a theory of abundance, distribution, and energetics. Oxford University Press, Oxford, UK.
  
}
%~~ Optionally other standard keywords, one per line, ~~
%~~ from file KEYWORDS in the R documentation directory ~~
% \keyword{ entropy  }
\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%~~ simple examples of the most important functions ~~
}
