% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meter_detect_PTR.R
\name{meter_detect_PTR}
\alias{meter_detect_PTR}
\title{Computes 'proportions of tumor-like reads' PTR for each sample and creates a summary table}
\usage{
meter_detect_PTR(path_read_tables, min_sites = 6, ncores = 1)
}
\arguments{
\item{path_read_tables}{The absolute path to the folder containing all the Read Tables generated for each sample using the create_read_table function.}

\item{min_sites}{An integer specifying the minimum number of CpGs a read must contain to be included in the PTR calculation (default: min_sites = 6).}

\item{ncores}{An integer specifying the number of processor cores to use for parallel processing of samples (default: ncores = 1).}
}
\value{
A DataFrame ('PTR Table') containing PTR values for each sample, including `PTR_hyper` (for hypermethylated DMR), `PTR_hypo` (for hypomethylated DMR), and `PTR_all` (combining both hyper- and hypomethylated DMR). The `PTR_all` metric, which integrates information from both hyper- and hypomethylated DMR, is recommended for classifying samples as ctDNA+/-.
}
\description{
Computes 'proportions of tumor-like reads' PTR for each sample and creates a summary table
}
