% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meter_subtype.R
\name{meter_subtype}
\alias{meter_subtype}
\title{Infer tumor subtype from cfDNA samples}
\usage{
meter_subtype(ref_mat, dmr_beta)
}
\arguments{
\item{ref_mat}{A matrix or DataFrame of median beta-values for DMR (row names) for reference components (column names), including a column labeled `healthy_cfDNA` with median beta-values of the reference cfDNA component.}

\item{dmr_beta}{A DataFrame containing beta-values for DMR (row names) across cfDNA samples to be analyzed (column names).}
}
\value{
A DataFrame where each row corresponds to an analyzed sample, containing the proportions of each component and an additional `pred_subtype` column indicating the inferred subtype.
}
\description{
Infer tumor subtype from cfDNA samples
}
