% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sar.R
\name{meteSAR}
\alias{meteSAR}
\title{Compute METE species area relationship (SAR)}
\usage{
meteSAR(spp, abund, row, col, x, y, S0 = NULL, N0 = NULL, Amin, A0,
  upscale = FALSE, EAR = FALSE)
}
\arguments{
\item{spp}{}

\item{row}{}

\item{col}{}

\item{x}{}

\item{y}{}

\item{S0}{}

\item{N0}{}

\item{Amin}{}

\item{A0}{}

\item{upscale}{}

\item{EAR}{}

\item{abun}{}
}
\description{
Compute METE species area relationship (SAR)
}
\examples{
esf1=meteESF(spp=arth$spp,
              abund=arth$count,
              power=arth$mass^(.75),
              minE=min(arth$mass^(.75)))
}
\author{
Andy Rominger <ajrominger@gmail.com>, Cory Merow
}
\references{
Harte, J. 2011. Maximum entropy and ecology: a theory of abundance, distribution, and energetics. Oxford University Press.
}

