% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/residuals.meteRelat.R
\name{mse.meteRelat}
\alias{mse.meteRelat}
\title{Computes mean squared error for meteRelat}
\usage{
\method{mse}{meteRelat}(x)
}
\arguments{
\item{x}{a \code{meteRelat} object}
}
\value{
numeric; the value of the mean squared error.
}
\description{
\code{mse.meteRelat} computes mean squared error for rank or cdf between METE prediction and data
}
\details{
See Examples.
}
\examples{
esf1=meteESF(spp=arth$spp,
              abund=arth$count,
              power=arth$mass^(.75),
              minE=min(arth$mass^(.75)))
sad1=sad.meteESF(esf1)
mse.meteDist(sad1, type='rank', relative=FALSE)
}
\author{
Andy Rominger <ajrominger@gmail.com>, Cory Merow
}
\references{
Harte, J. 2011. Maximum entropy and ecology: a theory of abundance, distribution, and energetics. Oxford University Press.
}
\seealso{
mseZ.meteDist
}

