\name{plot.mete}
\alias{plot.mete}
\title{Title of function}
\usage{
plot.mete(x, type = c("abundance", "power"), pers = c("rank", "cumulative"),
  ...)
}
\arguments{
  \item{arg}{description of arg}

  \item{arg}{description of arg}
}
\value{
- the type of object that the function returns
}
\description{
\code{function.name} what it does
}
\details{
how it works etc.
}
\examples{
#code to run
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
\keyword{manip}

