% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.meteDist}
\alias{plot.meteDist}
\title{Plot to compare METE predictions to data}
\usage{
plot.meteDist(x, ptype = c("cdf", "rad"), th.col = "red",
  lower.tail = TRUE, add.legend = TRUE, ...)
}
\arguments{
\item{x}{a \code{meteDist} object}

\item{ptype}{either "cdf" or "rad"}

\item{th.col}{line color of theoretical prediction}

\item{lower.tail}{logical; choose TRUE to highlight differences between data and theory at low abundance; choose FALSE to highlight differences at high abundance.}

\item{add.legend}{logical; add a legend}

\item{...}{arguments to be passed to \code{plot}}
}
\description{
\code{plot.meteDist} plots both the theoretical prediction and data for a \code{meteDist} object using either a rank or cumulative distribution plot
}
\details{
\code{plot.meteDist} automatically extracts the prediction and data from the \code{meteDist} objects. Additional plotting arguments can be passed from \code{par}.
}
\examples{
esf1=meteESF(spp=arth$spp,
              abund=arth$count,
              power=arth$mass^(.75),
              minE=min(arth$mass^(.75)))
ipd1=ipd.meteESF(esf1)
plot(ipd1)
plot(ipd1, ptype='rad')
}
\author{
Andy Rominger <ajrominger@gmail.com>, Cory Merow
}
\references{
Harte, J. 2011. Maximum entropy and ecology: a theory of abundance, distribution, and energetics. Oxford University Press.
}

