% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.meteRelat.R
\name{plot.meteRelat}
\alias{plot.meteRelat}
\title{plot the ???}
\usage{
\method{plot}{meteRelat}(x, legend = TRUE, ...)
}
\arguments{
\item{x}{}

\item{legend}{Logical}

\item{...}{arguments to be passed to plot}
}
\description{
plot the ???
}
\examples{
esf1=meteESF(spp=arth$spp,
              abund=arth$count,
              power=arth$mass^(.75),
              minE=min(arth$mass^(.75)))
}
\author{
Andy Rominger <ajrominger@gmail.com>, Cory Merow
}
\references{
Harte, J. 2011. Maximum entropy and ecology: a theory of abundance, distribution, and energetics. Oxford University Press.
}

