% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.sar.R
\name{plot.sar}
\alias{plot.sar}
\title{Plot the species abundance distribution (SAR), i.e. objects of class sar}
\usage{
\method{plot}{sar}(x, add = FALSE, ...)
}
\arguments{
\item{x}{an object of class SAR made with}
}
\description{
Plot the species abundance distribution (SAR), i.e. objects of class sar
}
\examples{
esf1=meteESF(spp=arth$spp,
              abund=arth$count,
              power=arth$mass^(.75),
              minE=min(arth$mass^(.75)))
}
\author{
Andy Rominger <ajrominger@gmail.com>, Cory Merow
}
\references{
Harte, J. 2011. Maximum entropy and ecology: a theory of abundance, distribution, and energetics. Oxford University Press.
}

