\name{predict.meteDist}
\alias{predict.meteDist}
\title{Title of function}
\usage{
\method{predict}{meteDist}(x, rank = TRUE, ...)
}
\arguments{
  \item{arg}{description of arg}

  \item{arg}{description of arg}
}
\value{
- the type of object that the function returns
}
\description{
\code{function.name} what it does
}
\details{
how it works etc.
}
\examples{
#code to run
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
\keyword{manip}

