% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.metePi}
\alias{print.metePi}
\title{print.metePi}
\usage{
\method{print}{metePi}(x, ...)
}
\arguments{
\item{x}{an object of class \code{metePi}}

\item{...}{arguments to be passed}
}
\description{
\code{print.metePi} prints an object of class \code{metePi}
}
\details{
See Examples
}
\author{
Andy Rominger <ajrominger@gmail.com>, Cory Merow
}
\references{
Harte, J. 2011. Maximum entropy and ecology: a theory of abundance, distribution, and energetics. Oxford University Press.
}
\keyword{METE,}
\keyword{MaxEnt,}
\keyword{ecosystem}
\keyword{function}
\keyword{lagrange}
\keyword{multiplier,}
\keyword{structure}

