% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.sar.R
\name{print.sar}
\alias{print.sar}
\title{print.meteESF}
\usage{
\method{print}{sar}(x)
}
\arguments{
\item{x}{an object of class \code{meteESF}}

\item{...}{arguments to be passed}
}
\value{
Returns the object silently
}
\description{
\code{print.meteESF} prints an object of class \code{meteESF}
}
\details{
See Examples
}
\examples{
data(anbo)
anbo.sar <- meteSAR(anbo.new$spp, anbo.new$count, anbo.new$row, anbo.new$col, Amin=1, A0=16)
print(anbo.sar)
anbo.sar # alternatively
}
\author{
Andy Rominger <ajrominger@gmail.com>, Cory Merow
}
\keyword{METE,}
\keyword{MaxEnt,}
\keyword{ecosystem}
\keyword{function}
\keyword{lagrange}
\keyword{multiplier,}
\keyword{structure}

