% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sad.meteESF.R
\name{sad}
\alias{sad}
\title{generic method for METE species abundance distribution}
\usage{
sad(x)
}
\arguments{
\item{esf}{an object of class mete.}
}
\value{
An object of class \code{meteDist} which inherits from objects of class \code{sad}. The object contains a list with the following elements.
\describe{
   \item{\code{type}}{'sad'}
   \item{\code{data}}{X}
   \item{\code{d}}{this.eq}
   \item{\code{p}}{FUNp}
   \item{\code{q}}{FUNq}
   \item{\code{r}}{FUNr}
}
}
\description{
\code{sad.mete} returns the species abundance distribution
predicted by METE (Phi(n))
}
\details{
See Examples.
}
\examples{
esf1=meteESF(spp=arth$spp,
              abund=arth$count,
              power=arth$mass^(.75),
              minE=min(arth$mass^(.75)))
sad=sad(esf1)
}
\author{
Andy Rominger <ajrominger@gmail.com>, Cory Merow
}
\references{
Harte, J. 2011. Maximum entropy and ecology: a theory of abundance, distribution, and energetics. Oxford University Press.
}
\seealso{
metePhi
}
\keyword{METE,}
\keyword{MaxEnt,}
\keyword{ecosystem}
\keyword{function}
\keyword{lagrange}
\keyword{multiplier,}
\keyword{structure}

