% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{serp}
\alias{serp}
\title{Community abundance data for a serpentine grassland}
\format{A data frame with 1928 rows and 4 variables:
\describe{
  \item{row}{plot coordinate}
  \item{column}{plot coordinate}
  \item{spp}{species ID}
  \item{count}{number of individuals}
}}
\source{
Green J., J. Harte, and A. Ostling, 2003. Species richness, endemism, and abundance patterns: tests of two fractal models in a serpentine grassland. Ecology Letters 6: 919--928.
}
\usage{
serp
}
\description{
A dataset containing the community abundace data for plant species, as well
as the locations of plots with respect to one another
}
\keyword{datasets}

