% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ssad.metePi.R
\name{ssad.metePi}
\alias{ssad.metePi}
\title{Species Spatial Abundance Distribution}
\usage{
ssad.metePi(ssf)
}
\arguments{
\item{ssf}{}
}
\description{
Species Spatial Abundance Distribution
}
\examples{
esf1=meteESF(spp=arth$spp,
              abund=arth$count,
              power=arth$mass^(.75),
              minE=min(arth$mass^(.75)))
}
\author{
Andy Rominger <ajrominger@gmail.com>, Cory Merow
}
\references{
Harte, J. 2011. Maximum entropy and ecology: a theory of abundance, distribution, and energetics. Oxford University Press.
}

