% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sar.R
\name{upscaleSAR}
\alias{upscaleSAR}
\title{upscale SAR}
\usage{
upscaleSAR(x, A0, Aup, EAR = FALSE)
}
\arguments{
\item{x}{}

\item{A0}{}

\item{Aup}{}

\item{EAR}{logical. TRUE computes the endemics-area relatinship}
}
\value{
list
}
\description{
upscale SAR
}
\examples{
esf=meteESF(spp=anbo$spp,
             abund=anbo$count)
}
\author{
Andy Rominger <ajrominger@gmail.com>, Cory Merow
}
\references{
Harte, J. 2011. Maximum entropy and ecology: a theory of abundance, distribution, and energetics. Oxford University Press.
}

