% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{arth}
\alias{arth}
\title{Arthropod community abundance data}
\format{A data frame with 547 rows and 3 variables:
\describe{
  \item{spp}{species ID}
  \item{count}{number of individuals}
  \item{mass}{biomass}
}}
\source{
Gruner, D. S. 2007. Geological age, ecosystem development, and local resource constraints on arthropod community structure in the Hawaiian Islands. Biological Journal of the Linnean Society, 90: 551--570.
}
\usage{
arth
}
\description{
A dataset containing the community abundace data for individuals, as well 
as their body mass.
}
\keyword{datasets}

