% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dms_beta_table.R
\name{create_dms_beta_table}
\alias{create_dms_beta_table}
\title{Create table of DMS beta-values}
\usage{
create_dms_beta_table(dms_table, path_cov_files, id_pattern = NULL)
}
\arguments{
\item{dms_table}{A DataFrame specifying the chromosomal and genomic positions of selected DMS, with at least 4 required columns: `dms_id` (unique identifier for each DMS), `chr` (chromosome, e.g., 'chr1', 'chr2'), `pos` (genomic position), and `type` (methylation status, either 'hypo' for hypomethylated or 'hyper' for hypermethylated DMS)}

\item{path_cov_files}{The absolute path to the folder containing the filtered coverage files of samples to be analyzed, which should have been created using the `METER::filter_cov_alpha100` function.}

\item{id_pattern}{A string to be used as input for the R `strsplit()` function to extract sample names from the base names of the input "coverage files." If not specified (default = NULL), the sample names will be directly obtained from the base names of the input "coverage files".}
}
\value{
A DataFrame (Beta Table) containing the beta-values (proportions) of the selected DMS, with DMS as row names (identified by their unique `dms_id`, consistent with the `dms_table`) and samples as column names. Beta-values are calculated based on reads with alpha value of 1 and that cover a minimum number of CpGs.
}
\description{
Create table of DMS beta-values
}
