% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_cov_alpha100.R
\name{filter_cov_alpha100}
\alias{filter_cov_alpha100}
\title{Filter Bismark cpg files considering only reads with alpha=100}
\usage{
filter_cov_alpha100(
  path_bismark2bedGraph,
  path_cpg_file,
  path_read_table,
  path_out = NULL,
  min_sites = 6,
  remove_cpg = FALSE
)
}
\description{
Filter reads using reads-dmr table created in previous step
}
\details{
input:
for each sample:
- Bismark output "CpG file"
- reads-dmr summary table with specific columns: read_id, meth_perc (alpha of read, i.e. degree of methylation of read)...
- minimum number of CpG sites that reads with alpha=100 must cover
- path to output folder (where to save the filtered cpg file)

output:
for each sample:
Bismark-like cpg context file comprising reads with alpha=100 only
}
