% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ipd.meteESF.R
\name{ipd.meteESF}
\alias{-}
\alias{a}
\alias{additional}
\alias{be}
\alias{command}
\alias{documentation}
\alias{from}
\alias{ipd.meteESF}
\alias{line.}
\alias{list}
\alias{looks}
\alias{mapped}
\alias{names}
\alias{of}
\alias{that}
\alias{the}
\alias{them}
\alias{this}
\alias{to}
\alias{topic}
\alias{up}
\alias{user}
\alias{when}
\alias{will}
\title{Individual Power Distribution}
\usage{
\method{ipd}{meteESF}(esf)
}
\arguments{
\item{esf}{an object of class mete.}
}
\value{
list
}
\description{
\code{ipd.meteESF} calculates the distribution Psi(e | N0, S0, E0) (the distribution of metabolic rates across all individuals in a commmunity)
}
\details{
See examples.
}
\examples{
esf1=meteESF(spp=arth$spp,
              abund=arth$count,
              power=arth$mass^(.75),
              minE=min(arth$mass^(.75)))
ipd1=ipd(esf1)
}
\author{
Andy Rominger <ajrominger@gmail.com>, Cory Merow
}
\references{
Harte, J. 2011. Maximum entropy and ecology: a theory of abundance, distribution, and energetics. Oxford University Press.
}
\seealso{
sad.meteESF, metePsi
}
\keyword{METE,}
\keyword{MaxEnt,}
\keyword{ecosystem}
\keyword{function}
\keyword{lagrange}
\keyword{multiplier,}
\keyword{structure}

