% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metePi.R
\name{metePi}
\alias{metePi}
\title{make Pi distribution}
\usage{
metePi(abund, n0 = sum(abund), A, A0)
}
\arguments{
\item{abund}{}

\item{n0}{}

\item{A}{}

\item{A0}{}
}
\description{
make Pi distribution
}
\examples{
esf1=meteESF(spp=arth$spp,
              abund=arth$count,
              power=arth$mass^(.75),
              minE=min(arth$mass^(.75)))
}
\author{
Andy Rominger <ajrominger@gmail.com>, Cory Merow
}
\references{
Harte, J. 2011. Maximum entropy and ecology: a theory of abundance, distribution, and energetics. Oxford University Press.
}

