% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meter_detect_PTR.R
\name{meter_detect_PTR}
\alias{meter_detect_PTR}
\title{Compute PTR of a sample considering only reads with alpha=100
Create PTR table for all study samples}
\usage{
meter_detect_PTR(path_read_tables, min_sites = 6, ncores = 1)
}
\description{
input:
- dmr table with specific columns: dmr_id, type (hyper/hypo)
- reads-dmr summary table with specific columns: read_id, dmr_id, meth_perc (alpha of read, i.e. degree of methylation of read)
}
\details{
NB: dmr_id in the reads-dmr table sholud be a subset of the dmr_id in the dmr table

output: data.frame with 1 row containing sample_id, ptr hyper, ptr hypo and ptr glob
}
