% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meter_quant_PTS.R
\name{meter_quant_PTS}
\alias{meter_quant_PTS}
\title{Create PTS table}
\usage{
meter_quant_PTS(dms_table, beta_table)
}
\description{
input:
- dms table with specific columns: dms_id, type (hyper/hypo)
- dms beta table obtained from filtered* cov files
}
\details{
*filtered cov files: cov files obtained considering only reads with alpha=100

output: data.frame with PTS hyper/hypo/glob (columns) for each sample (rows)
}
