% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/meteSSF.R
\name{meteSSF}
\alias{meteSSF}
\title{meteSSF}
\usage{
meteSSF(abund, n0 = sum(abund), A, A0)
}
\arguments{
\item{abund}{A vector giving abundances of each entry}

\item{n0}{Total abundance in area A0}

\item{A}{The area at which abundances were recorded}

\item{A0}{Total study area}
}
\value{
An object of class \code{meteSSF} with elements
\describe{
  \item{\code{data}}{The data used to construct the SSF}
  \item{\code{La}}{Vector of Lagrange multipliers}
  \item{\code{La.info}}{Termination information from optimization procedure}
  \item{\code{state.var}}{State variables used to constrain entropy maximization}
}
}
\description{
\code{meteSSF} calculates the ``spatial structure
function'' \eqn{\Pi(n)} (analygous to the ecosystem structure function). From
the SSF the spatial abundance distribution can be calculated.
}
\details{
Uses either data or state variables to calculate the Spatial Structure
Function (SSF). Uses internal code to determine when computation-saving approximations
can be safely made
}
\examples{
data(anbo)

## note: abundance of each spp needs to be recorded in each cell
## even if that abundance is 0....so might need to do that in some
## data sets

samp2mat <- function(site,spp,abund) {
  y <- tapply(abund, list(site, spp), sum)
  y[is.na(y)] <- 0
  return(y)
}

anbo.mat <- samp2mat(paste(anbo[, 1], anbo[,2]), anbo$spp, anbo$count)
anbo.new <- data.frame(t(sapply(strsplit(rownames(anbo.mat), ' ', fixed=TRUE), as.numeric)),
                       spp = rep(colnames(anbo.mat), each=nrow(anbo.mat)),
                       count = as.vector(anbo.mat))
colnames(anbo.new)[1:2] <- colnames(anbo)[1:2]

## anbo.new now has 0 abundance where needed

## calculate SSF Pi
pi1 <- meteSSF(anbo.new$count[anbo.new$spp=='crcr'], A=1, A0=16)
pi1
}
\author{
Andy Rominger <ajrominger@gmail.com>, Cory Merow
}
\references{
Harte, J. 2011. Maximum entropy and ecology: a theory of abundance, distribution, and energetics. Oxford University Press.
}
\seealso{
metePi
}
\keyword{METE,}
\keyword{MaxEnt}
\keyword{lagrange}
\keyword{multiplier,}

