% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sipd.meteDist.R
\name{meteTheta}
\alias{meteTheta}
\title{Title of function}
\usage{
meteTheta(e, n, la2)
}
\arguments{
\item{arg}{description of arg}

\item{arg}{description of arg}
}
\value{
- the type of object that the function returns
}
\description{
\code{function.name} what it does
}
\details{
how it works
etc.
}
\examples{
#code to run
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
\keyword{manip}

