% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sipd.meteDist.R
\name{meteTheta}
\alias{meteTheta}
\title{Intra-specific metabolic rate distribution}
\usage{
meteTheta(e, n, la2)
}
\arguments{
\item{e}{}

\item{n}{}

\item{la2}{}
}
\description{
Distribu􏴜on of metabolic rates over individuals within a species of abundance n0 (Theta)
# $\Theta (\epsilon | n0,S0,N0,E0) = \dfrac{R}{\phi}$
#
}
\details{
\
}
\examples{
esf1=meteESF(spp=arth$spp,
              abund=arth$count,
              power=arth$mass^(.75),
              minE=min(arth$mass^(.75)))
}
\author{
Andy Rominger <ajrominger@gmail.com>, Cory Merow
}
\references{
Harte, J. 2011. Maximum entropy and ecology: a theory of abundance, distribution, and energetics. Oxford University Press.
}

