\name{meteZ}
\alias{meteZ}
\title{Title of function}
\usage{
meteZ(la1, la2, S0, N0, E0)

meteZ(la1, la2, S0, N0, E0)
}
\arguments{
  \item{arg}{description of arg}

  \item{arg}{description of arg}

  \item{la1}{description of arg}

  \item{la2}{description of arg}

  \item{S0}{description of arg}

  \item{N0}{description of arg}

  \item{E0}{description of arg}
}
\value{
- the type of object that the function returns

- the type of object that the function returns
}
\description{
\code{function.name} what it does

function to return Z, the normalizing constant of R as well
as the simplifying parameters beta and sigma
\code{function.name} what it does
}
\details{
how it works etc.

how it works etc.
}
\examples{
#code to run
#code to run
}
\author{
Andy Rominger <ajrominger@gmail.com>

Andy Rominger <ajrominger@gmail.com>
}
\keyword{manip}

