% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.damuth.R
\name{print.damuth}
\alias{print.damuth}
\title{print.damuth}
\usage{
\method{print}{damuth}(x, ...)
}
\arguments{
\item{x}{an object of class \code{damuth}}

\item{...}{arguments to be passed to methods}
}
\value{
Returns the object silently
}
\description{
S3 method for class \code{damuth}
}
\details{
See Examples
}
\examples{
data(arth)
esf1 <- meteESF(arth$spp, arth$count, arth$mass^0.75)
ebar1 <- ebar(esf1)
print(ebar1)

}
\author{
Andy Rominger <ajrominger@gmail.com>, Cory Merow
}

