% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.meteRelat.R
\name{print.meteRelat}
\alias{print.meteRelat}
\title{Print summaries of \code{meteRelat} objects}
\usage{
\method{print}{meteRelat}(x)
}
\arguments{
\item{x}{}
}
\value{
x silently
}
\description{
S3 method for class \code{meteRelat}
}
\author{
Andy Rominger <ajrominger@gmail.com>, Cory Merow
}
\references{
Harte, J. 2011. Maximum entropy and ecology: a theory of abundance, distribution, and energetics. Oxford University Press.
}

