\name{residuals.meteDist}
\alias{residuals.meteDist}
\title{Title of function}
\usage{
\method{residuals}{meteDist}(X, type = c("rank", "cumulative"),
  relative = TRUE, log = FALSE)
}
\arguments{
  \item{arg}{description of arg}

  \item{arg}{description of arg}
}
\value{
- the type of object that the function returns
}
\description{
\code{function.name} what it does
}
\details{
how it works etc.
}
\examples{
#code to run
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
\keyword{manip}

