% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sipd.meteDist.R
\name{sipd}
\alias{sipd}
\title{Generic method to obtain the species-level individual power distribution (SIPD)}
\usage{
sipd(x, ...)
}
\arguments{
\item{esf}{}

\item{sppID}{}

\item{n}{}
}
\description{
Extract species level individual power distribution
from ESF object and return object inheriting from meteDist
}
\examples{
esf1=meteESF(spp=arth$spp,
              abund=arth$count,
              power=arth$mass^(.75),
              minE=min(arth$mass^(.75)))
sipd1 <- sipd(arth.esf, sppID=5)
}
\author{
Andy Rominger <ajrominger@gmail.com>, Cory Merow
}
\references{
Harte, J. 2011. Maximum entropy and ecology: a theory of abundance, distribution, and energetics. Oxford University Press.
}

