% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ssad.metePi.R
\name{ssad}
\alias{ssad}
\alias{ssad.meteSSF}
\title{Species Spatial Abundance Distribution}
\usage{
ssad(x)

\method{ssad}{meteSSF}(ssf)
}
\arguments{
\item{ssf}{An objects of class meteSSF; i.e. the spatial structure function \eqn{\Pi(n)}}
}
\description{
Species Spatial Abundance Distribution
}
\examples{
data(anbo)
samp2mat <- function(site,spp,abund) {
    y <- tapply(abund, list(site, spp), sum)
    y[is.na(y)] <- 0
    return(y)
}
anbo.mat <- samp2mat(paste(anbo[, 1], anbo[,2]), anbo$spp, anbo$count)
anbo.new <- data.frame(t(sapply(strsplit(rownames(anbo.mat), ' ', fixed=TRUE), as.numeric)), spp = rep(colnames(anbo.mat), each=nrow(anbo.mat)), count = as.vector(anbo.mat))
colnames(anbo.new)[1:2] <- colnames(anbo)[1:2]
## anbo.new now has 0 abundance where needed
pi1 <- meteSSF(anbo.new$count[anbo.new$spp=='crcr'], A=1, A0=16)
plot(ssad(pi1))
}
\author{
Andy Rominger <ajrominger@gmail.com>, Cory Merow
}
\references{
Harte, J. 2011. Maximum entropy and ecology: a theory of abundance, distribution, and energetics. Oxford University Press.
}

